move containers to container folders
